Configure onLibsInit and getWeb3Config
